+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+- Use composer-based console scripts.
+
 ## 2.1.2 - 2025-01-11
 
 ### Fixed